// types/firefly.d.ts
export interface paths {
// ...
    "/v1/transactions": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List all the user's transactions.
         * @description List all the user's transactions.
         */
        get: operations["listTransaction"];
        put?: never;
        /**
         * Store a new transaction
         * @description Creates a new transaction. The data required can be submitted as a JSON body or as a list of parameters.
         */
        post: operations["storeTransaction"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Transaction: {
            /**
             * Format: date-time
             * @example 2025-11-01T00:00:00+00:00
             */
            readonly created_at?: string;
            /**
             * Format: date-time
             * @example 2025-11-01T00:00:00+00:00
             */
            readonly updated_at?: string;
            /**
             * Format: string
             * @description User ID
             * @example 3
             */
            readonly user?: string;
            /**
             * Format: string
             * @description Title of the transaction if it has been split in more than one piece. Empty otherwise.
             * @example Split transaction title.
             */
            group_title?: string | null;
            transactions: components["schemas"]["TransactionSplit"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}